+++ Sources/Navigator/PDF/PDFNavigatorViewController.swift
@@ -12,12 +12,15 @@ import R2Shared
 
 public protocol PDFNavigatorDelegate: VisualNavigatorDelegate, SelectableNavigatorDelegate { }
 
-
 /// A view controller used to render a PDF `Publication`.
 @available(iOS 11.0, *)
 open class PDFNavigatorViewController: UIViewController, VisualNavigator, SelectableNavigator, Loggable {
     
     enum Error: Swift.Error {
+        /// The provided publication is restricted. Check that any DRM was
+        /// properly unlocked using a Content Protection.
+        case publicationRestricted
+
         case openPDFFailed
     }
     
@@ -39,11 +42,75 @@ open class PDFNavigatorViewController: UIViewController, VisualNavigator, Select
     /// Holds a reference to make sure it is not garbage-collected.
     private var tapGestureController: PDFTapGestureController?
 
-    public init(publication: Publication, initialLocation: Locator? = nil, editingActions: [EditingAction] = EditingAction.defaultActions) {
-        assert(!publication.isRestricted, "The provided publication is restricted. Check that any DRM was properly unlocked using a Content Protection.")
+    private let server: HTTPServer?
+    private let publicationEndpoint: HTTPServerEndpoint?
+    private let publicationBaseURL: URL
+
+    public convenience init(
+        publication: Publication,
+        initialLocation: Locator?,
+        editingActions: [EditingAction] = EditingAction.defaultActions,
+        httpServer: HTTPServer
+    ) throws {
+        guard !publication.isRestricted else {
+            throw Error.publicationRestricted
+        }
+
+        let publicationEndpoint: HTTPServerEndpoint?
+        let baseURL: URL
+        if let url = publication.baseURL {
+            publicationEndpoint = nil
+            baseURL = url
+        } else {
+            let endpoint = UUID().uuidString
+            publicationEndpoint = endpoint
+            baseURL = try httpServer.serve(at: endpoint, publication: publication)
+        }
+
+        self.init(
+            publication: publication,
+            initialLocation: initialLocation,
+            httpServer: httpServer,
+            publicationEndpoint: publicationEndpoint,
+            publicationBaseURL: baseURL,
+            editingActions: editingActions
+        )
+    }
+
+    @available(*, deprecated, message: "See the 2.5.0 migration guide to migrate the HTTP server")
+    public convenience init(
+        publication: Publication,
+        initialLocation: Locator? = nil,
+        editingActions: [EditingAction] = EditingAction.defaultActions
+    ) {
+        precondition(!publication.isRestricted, "The provided publication is restricted. Check that any DRM was properly unlocked using a Content Protection.")
+        guard let baseURL = publication.baseURL else {
+            preconditionFailure("No base URL provided for the publication. Add it to the HTTP server.")
+        }
         
+        self.init(
+            publication: publication,
+            initialLocation: initialLocation,
+            httpServer: nil,
+            publicationEndpoint: nil,
+            publicationBaseURL: baseURL,
+            editingActions: editingActions
+        )
+    }
+
+    private init(
+        publication: Publication,
+        initialLocation: Locator?,
+        httpServer: HTTPServer?,
+        publicationEndpoint: HTTPServerEndpoint?,
+        publicationBaseURL: URL,
+        editingActions: [EditingAction]
+    ) {
         self.publication = publication
         self.initialLocation = initialLocation
+        self.server = httpServer
+        self.publicationEndpoint = publicationEndpoint
+        self.publicationBaseURL = URL(string: publicationBaseURL.absoluteString.addingSuffix("/"))!
         self.editingActions = EditingActionsController(actions: editingActions, rights: publication.rights)
         
         super.init(nibName: nil, bundle: nil)
@@ -66,6 +133,10 @@ open class PDFNavigatorViewController: UIViewController, VisualNavigator, Select
     
     deinit {
         NotificationCenter.default.removeObserver(self)
+
+        if let endpoint = publicationEndpoint {
+            server?.remove(at: endpoint)
+        }
     }
 
     open override func viewDidLoad() {
@@ -208,7 +279,7 @@ open class PDFNavigatorViewController: UIViewController, VisualNavigator, Select
         }
         
         if currentResourceIndex != index {
-            guard let url = link.url(relativeTo: publication.baseURL),
+            guard let url = link.url(relativeTo: publicationBaseURL),
                 let document = PDFDocument(url: url) else
             {
                 log(.error, "Can't open PDF document at \(link)")
@@ -421,21 +492,6 @@ extension PDFNavigatorViewController: UIGestureRecognizerDelegate {
 
     public func gestureRecognizer(_ gestureRecognizer: UIGestureRecognizer, shouldRecognizeSimultaneouslyWith otherGestureRecognizer: UIGestureRecognizer) -> Bool {
         return true
-    }
-    
-}
-
-
-// MARK: - Deprecated
-
-@available(iOS 11.0, *)
-extension PDFNavigatorViewController {
-    
-    /// This initializer is deprecated.
-    /// `license` is not needed anymore.
-    @available(*, unavailable, renamed: "init(publication:initialLocation:editingActions:)")
-    public convenience init(publication: Publication, license: DRMLicense?, initialLocation: Locator? = nil, editingActions: [EditingAction] = EditingAction.defaultActions) {
-        self.init(publication: publication, initialLocation: initialLocation, editingActions: editingActions)
     }
     
 }