+++ TestApp/Sources/Reader/CBZ/CBZModule.swift
@@ -27,8 +27,8 @@ final class CBZModule: ReaderFormatModule {
         return publication.conforms(to: .divina)
     }
     
-    func makeReaderViewController(for publication: Publication, locator: Locator?, bookId: Book.Id, books: BookRepository, bookmarks: BookmarkRepository, highlights: HighlightRepository, resourcesServer: ResourcesServer) throws -> UIViewController {
-        let cbzVC = CBZViewController(publication: publication, locator: locator, bookId: bookId, books: books, bookmarks: bookmarks)
+    func makeReaderViewController(for publication: Publication, locator: Locator?, bookId: Book.Id, books: BookRepository, bookmarks: BookmarkRepository, highlights: HighlightRepository) throws -> UIViewController {
+        let cbzVC = try CBZViewController(publication: publication, locator: locator, bookId: bookId, books: books, bookmarks: bookmarks)
         cbzVC.moduleDelegate = self.delegate
         return cbzVC
     }